Remove .NET version from artifact path
The build output path for .NET includes the .NET version. This results in us having to update both the serverless config and build commands for Linux & Windows each time we update .NET.

This change removes the version from the path. This _should_ be safe because we only build for one version, and the three locations have been updated to be consistent._